64-bit serialize
Allow to create bigger than 4GB savestate/replay files. Likely we won't ever
create such huge files. But like many of the previous patches in this series,
this actually allows to slightly simplify the code by removing some explicit
truncating casts.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13090 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/config/XMLElement.hh
@@ -85,7 +85,7 @@ public:
 		string_ref name, string_ref attName,
 		string_ref attValue);
 	const XMLElement* findNextChild(string_ref name,
-	                                unsigned& fromIndex) const;
+	                                size_t& fromIndex) const;
 
 	std::vector<XMLElement*> getChildren(string_ref name) const;
